Move the host listener for changing themes to top level component for proper handling on Edge
